introduce Parser class, with two implementations
The Parser subsumes all the parsing logic. We have one instance
attribute which informs the parser if it should parse/output HTML for
our feeds. We also correct the circular interaction between parser.js
and manuscript.pegjs by passing the parser to be used as an option to
PEG when building the AST.